// packages/commercial-lytenyte-pro/src/types/api.ts
export interface DataRect {
  readonly rowStart: number;
  readonly rowEnd: number;
  readonly columnStart: number;
  readonly columnEnd: number;
}